+++ azplugins/BondPotentials.cu
@@ -1,0 +1,22 @@
+// Copyright (c) 2016-2017, Panagiotopoulos Group, Princeton University
+// This file is part of the azplugins project, released under the Modified BSD License.
+
+// Maintainer: astatt / Everyone is free to add additional potentials
+
+/*!
+ * \file BondPotentials.cu
+ * \brief Defines the driver functions for computing bonded forces on the GPU
+ *
+ * Each bond potential evaluator needs to have an explicit instantiation of the
+ * compute_bond_potential.
+ */
+
+#include "BondPotentials.cuh"
+
+namespace azplugins
+{
+namespace gpu
+{
+
+} // end namespace gpu
+} // end namespace azplugins