+++ azplugins/__init__.py
@@ -4,6 +4,7 @@
 # Maintainer: mphoward / All developers are free to add new modules
 
 import analyze
+import bond
 import evaporate
 import pair
 import update